// models/subcategory.ts
export interface SubCategory {
    id:                number;
    category_id:       number;
    name:              string;
    is_essential:      number;
    flexibility_score: number;
    spending_nature:   string;
    essentiality:      string;
    created_at:        null;
    updated_at:        null;
}